refactor: Move text nodes to individual files
Implements ML-597
